requirements.txt
numpy
Pillow
PyYAML
scikit-learn
joblib
scikit-image
tensorflow
streamlit
albumentations>=1.3.0
opencv-python>=4.8.0
